refactor(factory): move AliasNormalizingMiddleware into factory package
